Integrated all test cases in test.cpp
